Add files via upload
Hiya guys, this is what I have so far with the game (I have figured out some basic animation!). The code follows on from what the guy on YouTube does (for details and a decent description of the code, I'd very much recommend).

Trying to figure out how to make the rocket accelerate upwards and naturally (without keyboard presses) fall downwards due to gravity.

Note also that there are only two animation images per 'action'. i.e. when the rocket is flying, it chooses one file directory with two images and a different directory for when it's stationary (see the Coding with Russ video Soldier Game part 3). 

The animation is sooooo jarring because the images are out of scale, but if we keep a consistent image size (32x32 pixels) and have many more images, it will be far smoother. I think Lua and Hannah you can help me here by starting the drawings for the rocket and planets etc if you like!

Thanks guys (and sorry for the essay!),
Oscar